Probabilistic Forecast Combination - ProFoC

Returns predictions and weights calculated by online-learning algorithms using CRPS Learning. By default, the weights are calculated by gradient based bernstein online aggregation (BOAG).

profoc(y, experts, tau, loss_function = "quantile",
loss_parameter = 1, ex_post_smooth = FALSE, ex_post_fs = FALSE,
lambda = -Inf, method = "boa", method_var = "A", forget_regret = 0,
forget_performance = 0, fixed_share = 0, gamma = 1, ndiff = 1, deg = 3,
rel_nseg = 0.5, gradient = TRUE, loss_array = NULL, regret_array = NULL,
trace = TRUE, init_weights = NULL, lead_time = 0)

Arguments

y
A numeric vector of realizations.

experts
A an array of predictions with dimension (Observations, Quantiles, Experts).

tau
A numeric vector of probabilities. corresponding to the columns of experts.

loss_function
Either "quantile", "expectile" or "percentage".

loss_parameter
Optional parameter scaling the power of the loss function.

ex_post_smooth
Determines if smoothing is during or after online-learning. If true, contemporary weights are not affected but output weights are. If false (default) smoothed weights are also by the algorithm.

ex_post_fs
Analogous to ex_post_smooth: shall a fixed-share be added during (FALSE) or after online-learning (TRUE).

lambda
Penalization parameter used in the smoothing Step. -Inf causes the smoothing step to be skipped (default).

method
One of "boa", "ml_poly" or "ewa".

method_var
Allows to calculate slight variations of the BOA algorithm

forget_regret
Share of past regret not to be considered, resp. to be forgotten in every iteration of the algorithm. Defaults to 0.

forget_performance
Share of past performance not to be considered, resp. to be forgotten in every iteration of the algorithm when choosing the parameter combination. Defaults to 0.

fixed_share
Amount of fixed share to be added to the weights. Defaults to 0. 1 leads to uniform weights.

gamma
Scaling parameter for the learning rate.

ndiff
Degree of the differencing operator in the smoothing equation. 1 (default) leads to shrikage towards a constant.

deg
Degree of the B-Spine basis functions.

rel_nseg
determines how many basis functions are created in the smoothing step. This parameter defaults to 0.5 leading to 0.5*length(tau) to be used as number of knots. 1 Leads to as many knots as tau is long. 0 leads to one single knot.

gradient
Determines if a linearized version of the loss is used.

loss_array
User specified loss array. If specified, the loss will not be calculated by profoc.

regret_array
User specified regret array. If specifiec, the regret will not be calculated by profoc.

trace
If a progessbar shall be printed. Defaults to TRUE.

init_weights
Matrix of dimension Kx1 or KxP used as starting weights. Kx1 represents the constant solution with equal weights over all P whereas specifiying a KxP matrix allows different starting weights for each P.

lead_time
offset for expert forecasts. Defaults to 0, which means that experts forecast t+1 at t. Setting this to h means experts predictions refer to t+1+h at time t. The weight updates delay accordingly.

Value

Profoc can tune various parameters automatically based on the past loss. For this, lambda, forget, fixed_share, gamma, ndiff, deg and rel_nseg can be specified as numeric vectors containing parameters to consider. Profoc will automatically try all possible combinations of values provide.