Package Info

Use multiple online-aggregation algorithms to combine probabilistic forecasts using CRPS Learning as described in Berrisch, Ziel: "CRPS Learning", 2021. The primary function of this package is called online.

doi:10.1016/j.jeconom.2021.11.008

Details

Index: This package was not yet installed at build time.

Author

Maintainer: Jonathan Berrisch mailto:Jonathan@Berrisch.biz

Co-Author: Florian Ziel

References

Berrisch, Ziel: "CRPS Learning", 2021

doi:10.1016/j.jeconom.2021.11.008

doi:10.48550/arXiv.2102.00968

See also